+++ node/clients/AuthUser.ts
@@ -1,0 +1,17 @@
+import type { InstanceOptions, IOContext } from '@vtex/api'
+import { JanusClient } from '@vtex/api'
+
+export class AuthUser extends JanusClient {
+  constructor(ctx: IOContext, options?: InstanceOptions) {
+    super(ctx, {
+      ...options,
+      headers: {
+        ...options?.headers,
+      },
+    })
+  }
+
+  public async validateToken(token: any): Promise<any> {
+    return this.http.post(`/api/vtexid/credential/validate`, token)
+  }
+}